Handle decimals

+++ code/part03/language/src/main/java/toyl/ast/ToylArithmeticOpNode.java
@@ -15,7 +15,7 @@ public class ToylArithmeticOpNode extends ToylNode {
   }
 
   @Override
-  public long execute(VirtualFrame frame) {
+  public double execute(VirtualFrame frame) {
     var leftVal = this.left.execute(frame);
     var rightVal = this.right.execute(frame);
     return switch (this.binaryOp) {
